# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_fernandovalenzuela_api-home-pay_internal_models.Company:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      website_url:
        type: string
    type: object
...
  github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount:
    properties:
      account_identifier:
        type: string
      alias:
        type: string
      company:
        $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.Company'
      company_id:
        type: integer
      id:
        type: integer
    type: object
  github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
  internal_handlers.FinancialSummaryResponse:
    properties:
      balance:
        type: number
      expense_count:
        type: integer
      income_count:
        type: integer
      paid_expenses:
        type: number
      pending_expenses:
        type: number
      period:
        $ref: '#/definitions/internal_handlers.PeriodInfo'
      total_expenses:
        type: number
      total_incomes:
        type: number
    type: object
  internal_handlers.PeriodInfo:
    properties:
      id:
        type: integer
      month:
        type: integer
      year:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: API REST para gestión de gastos e ingresos del hogar
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: api-home-pay API
  version: "1.0"
paths:
...
  /api/service-accounts:
    get:
      consumes:
      - application/json
      description: Get all service accounts for the authenticated user, optionally
        filtered by company
      parameters:
      - description: Filter by company ID
        in: query
        name: company_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all service accounts
      tags:
      - service-accounts
    post:
      consumes:
      - application/json
      description: Create a new service account (e.g., "Electricity - Home")
      parameters:
      - description: Service account data
        in: body
        name: service_account
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create service account
      tags:
      - service-accounts
  /api/service-accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a service account by ID
      parameters:
      - description: Service Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete service account
      tags:
      - service-accounts
    get:
      consumes:
      - application/json
      description: Get a specific service account by ID
      parameters:
      - description: Service Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get service account by ID
      tags:
      - service-accounts
    put:
      consumes:
      - application/json
      description: Update an existing service account
      parameters:
      - description: Service Account ID
        in: path
        name: id
        required: true
        type: integer
      - description: Service account data
        in: body
        name: service_account
        required: true
        schema:
          $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_models.ServiceAccount'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update service account
      tags:
      - service-accounts
  /api/summary/{period_id}:
    get:
      consumes:
      - application/json
      description: Get financial summary for a specific period including incomes,
        expenses, balance
      parameters:
      - description: Period ID
        in: path
        name: period_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.FinancialSummaryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_fernandovalenzuela_api-home-pay_internal_utils.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get financial summary
      tags:
      - summary
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"